24/06/12 refactoring
