+++ Diary/Auth/AuthenticationViewModel.swift
@@ -148,15 +148,6 @@ extension AuthenticationViewModel {
   }
 }
 
-extension Date {
-  func isWithinPast(minutes: Int) -> Bool {
-    let now = Date.now
-    let timeAgo = Date.now.addingTimeInterval(-1 * TimeInterval(60 * minutes))
-    let range = timeAgo...now
-    return range.contains(self)
-  }
-}
-
 // MARK: - Sign in with Apple
 
 class SignInWithApple: NSObject, ASAuthorizationControllerDelegate {
