+++ Diary/Configurations/DateUtility.swift
@@ -6,3 +6,116 @@
 //
 
 import Foundation
+
+extension Date {
+    // Auth
+    func isWithinPast(minutes: Int) -> Bool {
+        let now = Date.now
+        let timeAgo = Date.now.addingTimeInterval(-1 * TimeInterval(60 * minutes))
+        let range = timeAgo...now
+        return range.contains(self)
+    }
+    
+    
+    
+    // This Month Start
+    func startOfMonth() -> Date {
+        let components = Calendar.current.dateComponents([.year, .month], from: self)
+        return Calendar.current.date(from: components)!
+    }
+    
+    func endOfMonth() -> Date {
+        let components:NSDateComponents = Calendar.current.dateComponents([.year, .month], from: self) as NSDateComponents
+        components.month += 1
+        components.day = 1
+        components.day -= 1
+        return Calendar.current.date(from: components as DateComponents)!
+    }
+    
+    // This Year Start
+    func startOfYeear() -> Date {
+        let components = Calendar.current.dateComponents([.year, .month], from: self)
+        return Calendar.current.date(from: components)!
+    }
+
+    func endOfYear() -> Date {
+        let interval = Calendar.current.dateInterval(of: .year, for: self)
+        return interval!.end
+    }
+    
+    // Convert UTC (or GMT) to local time
+    func toLocalTime() -> Date {
+        let timezone    = TimeZone.current
+        let seconds     = TimeInterval(timezone.secondsFromGMT(for: self))
+        return Date(timeInterval: seconds, since: self)
+    }
+    
+    
+    static var yesterday: Date {
+        return Date().dayBefore
+    }
+    
+    static var tomorrow:  Date {
+        return Date().dayAfter
+    }
+    var dayBefore: Date {
+        return Calendar.current.date(byAdding: .day, value: -1, to: noon)!
+    }
+    var dayAfter: Date {
+        return Calendar.current.date(byAdding: .day, value: 1, to: noon)!
+    }
+    var noon: Date {
+        return Calendar.current.date(bySettingHour: 1, minute: 0, second: 0, of: self)!
+    }
+    
+
+    func checkBoolDateIsWithinRange(date: Date) -> Bool {
+        
+        let calendar = Calendar.current
+        let components = calendar.dateComponents([.hour], from: Date.now)
+        let hour = components.hour ?? 0
+ 
+        var startDate = Date()
+        
+        if hour < 11 {
+            startDate = Calendar.current.date(byAdding: .day, value: -1, to: Date())!
+        }
+        
+        if Date().formatDate(startDate, using: .compare) == Date().formatDate(date, using: .compare) {
+            return true
+        } else {
+            return false
+        }
+    }
+    
+    func checkDateWithinRange(date: Date) -> Date {
+        let calendar = Calendar.current
+        let components = calendar.dateComponents([.hour], from: Date.now)
+        let hour = components.hour ?? 0
+        
+        var startDate = Date()
+        if hour < 11 {
+            startDate = Calendar.current.date(byAdding: .day, value: -1, to: Date())!
+        }
+
+        return startDate
+    }
+    
+    func formatDate(_ date: Date, using format: DateFormat_date) -> String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = format.rawValue
+        return formatter.string(from: date)
+    }
+}
+
+enum DateFormat_date: String {
+    case year = "YYYY"
+    case month = "MM"
+    case month_day = "MM / dd"
+    case month_text = "MM월"
+    case day = "dd"
+    case month_day_text = "MM월 dd일 일기"
+    case compare = "YYYYMMdd"
+}
+
+
