+++ Diary/Configurations/Utility.swift
@@ -6,3 +6,58 @@
 //
 
 import Foundation
+import SwiftUI
+
+extension View {
+  func endTextEditing() {
+    UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder),
+                                    to: nil, from: nil, for: nil)
+  }
+}
+
+extension Color {
+
+    init(hex: String) {
+        let scanner = Scanner(string: hex)
+        var rgbValue: UInt64 = 0
+        scanner.scanHexInt64(&rgbValue)
+
+        let r = (rgbValue & 0xff0000) >> 16
+        let g = (rgbValue & 0xff00) >> 8
+        let b = rgbValue & 0xff
+
+        self.init(red: Double(r) / 0xff, green: Double(g) / 0xff, blue: Double(b) / 0xff)
+    }
+}
+
+extension View {
+    
+    func changeOverlayOnScroll(
+        proxy : GeometryProxy,
+        offsetHolder : Binding<CGFloat>,
+        thresHold : Binding<CGFloat>,
+        toggle: Binding<Bool>
+    ) -> some View {
+        self
+            .onChange(
+                of: proxy.frame(in: .named("scroll")).minY
+            ) { newValue in
+                // Set current offset
+                offsetHolder.wrappedValue = abs(newValue)
+                // If current offset is going downward we hide overlay after 200 px.
+                if offsetHolder.wrappedValue > thresHold.wrappedValue + 200 {
+                    // We set thresh hold to current offset so we can remember on next iterations.
+                    thresHold.wrappedValue = offsetHolder.wrappedValue
+                    // Hide overlay
+                    toggle.wrappedValue = false
+                    
+                    // If current offset is going upward we show overlay again after 200 px
+                }else if offsetHolder.wrappedValue < thresHold.wrappedValue - 200 {
+                    // Save current offset to threshhold
+                    thresHold.wrappedValue = offsetHolder.wrappedValue
+                    // Show overlay
+                    toggle.wrappedValue = true
+                }
+         }
+    }
+}
