+++ Diary/View/MemoryView/MemoryDetailView.swift
@@ -83,15 +83,7 @@ struct MemoryDetailView: View {
                                 }
                             }.frame(height: CGFloat(image_height))
                                 .padding(.leading, AppConfig.homeWidth*0.08)
-//                                .onTapGesture {
-//                                    if image_ratio == 0.8 {
-//                                        image_ratio = 1.0
-//                                        image_height = 200
-//                                    } else {
-//                                        image_ratio = 0.8
-//                                        image_height = 150
-//                                    }
-//                                }
+
                         }
                         
                         HStack {
