+++ Diary/View/MemoryView/MemoryView.swift
@@ -30,8 +30,8 @@ struct MemoryView: View {
                         ForEach(group(todos, formatDate_Year_integer(_:)), id: \.self) { (section: [DB_core]) in
                             
                             NavigationLink(destination: {
-                                
-                                MemoryMonthView(section, titlestring: formatDate_Year(section[0].date!))
+                            
+                                MemoryMonthView(section, titlestring: Date().formatDate(section[0].date!, using: .year))
                                 
                             }, label: {
                                 //VStack으로 도형추가
@@ -60,7 +60,7 @@ struct MemoryView: View {
                                     VStack {
                                         Spacer()
                                         HStack {
-                                            Text(formatDate_Year(section[0].date!))
+                                            Text(Date().formatDate(section[0].date!, using: .year))
                                                 .foregroundStyle(.white)
                                                 .font(.custom("SUIT-Bold", size: 16))
                                             Spacer()
@@ -81,24 +81,6 @@ struct MemoryView: View {
         
     }
     
-    private func formatDate(_ date: Date) -> String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "YYYY / MM / dd"
-        return formatter.string(from: date)
-    }
-    
-    private func formatDate_Month(_ date: Date) -> String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "MM월"
-        return formatter.string(from: date)
-    }
-    
-    private func formatDate_Year(_ date: Date) -> String {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "YYYY"
-        return formatter.string(from: date)
-    }
-    
     private func formatDate_Year_integer(_ date: Date) -> Int {
         let formatter = DateFormatter()
         formatter.dateFormat = "YYYY"
@@ -113,52 +95,4 @@ struct MemoryView: View {
         }.values.map{ $0 }
     }
         
-}
-
-extension Date {
-    // This Month Start
-    func startOfMonth() -> Date {
-        let components = Calendar.current.dateComponents([.year, .month], from: self)
-        return Calendar.current.date(from: components)!
-    }
-    
-//    func startOfMonth() -> Date {
-//        let interval = Calendar.current.dateInterval(of: .month, for: self)
-//        return (interval?.start.toLocalTime())! // Without toLocalTime it give last months last date
-//    }
-    
-    func endOfMonth() -> Date {
-        let components:NSDateComponents = Calendar.current.dateComponents([.year, .month], from: self) as NSDateComponents
-        components.month += 1
-        components.day = 1
-        components.day -= 1
-        return Calendar.current.date(from: components as DateComponents)!
-    }
-    
-//    func endOfMonth() -> Date {
-//        let interval = Calendar.current.dateInterval(of: .month, for: self)
-//        return interval!.end
-//    }
-    
-    // This Year Start
-    func startOfYeear() -> Date {
-        let components = Calendar.current.dateComponents([.year, .month], from: self)
-        return Calendar.current.date(from: components)!
-    }
-//    func startOfYeear() -> Date {
-//        let interval = Calendar.current.dateInterval(of: .year, for: self)
-//        return (interval?.start.toLocalTime())!
-//    }
-//    
-    func endOfYear() -> Date {
-        let interval = Calendar.current.dateInterval(of: .year, for: self)
-        return interval!.end
-    }
-    
-    // Convert UTC (or GMT) to local time
-    func toLocalTime() -> Date {
-        let timezone    = TimeZone.current
-        let seconds     = TimeInterval(timezone.secondsFromGMT(for: self))
-        return Date(timeInterval: seconds, since: self)
-    }
 }
